fix: reading different media types and accessing iOS files after re-launching the app
On iOS container path change periodically (e.g. when app has been closed and launched again).
So instead of storing the absolute paths store relative path to the Directory.Documents and read files relative to that.
On top of that it should now scan subdirectories of the Documents folder, allowing files to be placed in different folders.

Additionally add a function to read MIME Type from file extensiomn to read different audio types properly.