feat(AuthorizationService): make AuthorizationService a generic service anything can implement
This was done mainly in preparation of implementing authentication for YouTube, however that is not worth the risk.
There is multiple sources documenting YouTube banning people that abuse downloading songs while being authenticated either via Cookies or OAuth.
Additionally YouTube's A/B testing of OAuth authentication makes it impossible to login for some accounts with it.

This is nonetheless a good change, I will most likely add AuthorizationService as a possible property on MusicPlayerService, so that these two won't be scattered.