feat: Rework object management
Completely reworks how objects such as Song, Album, Artist, etc. are stored and how they work.
Redundant and duplicate data was a big issue with how objects were stored to this point.
While this wouldn't be a big issue with small libraries, it would exponentially grow stored data, making it an issue for larger libraries.

Instead of storing duplicate data, e.g. 10 `Song`s in an `Album` could store the same `Artist`;
`Artist` is now allocated into a cache, and all of these songs store an `ArtistKey`, which is essentialy a pointer to the correct `Artist`. Same goes for other similar data.

The backing cache storage is currently just a simple object, however it will be changed to use a more organised system, since all of the data has identifying information such as `id`, `type` and `key`.