fix: Make cross-tab routes independant of library
To this point, sometimes routing would fail, e.g. after having a long navigation lists and then pressing tab button to go back.
After reading about ionic routing, it was caused by cross-tab routing, which would confuse the router, since it expects the history to be fully linear.
The workaround to that is to make routes, that can be accessed in the other routes independant of the others (so they have their own root route).

see: https://ionicframework.com/docs/vue/navigation#switching-between-tabs