feat: migrate from IndexedDB to files on a disk for `LocalImage`'s
Very large amount of images going through IndexedDB stresses it, which causes crashes due to its big size, weird behaviours and/or latency.
On top of that large amount of Blob URL's seems to cause issues where WKWebView might just restart with no specific reason to the user.
This move fixes all of those issues.

This commit migrates to the filesystem (both on Mobile and on Desktop), where browsers can directly access and manage images them however they please.
The only things left to do is deduplication and clearing them, where both are optional features as of now.